BC-76 | Moved schema, mandatory field, stage reached code, and UFN claim validation logic to separate validator classes to make testing easier.
